chore: Update dependency to discord.js 12.3.1

+++ src/commands/admin/prune.ts
@@ -1,5 +1,5 @@
 import { Argument } from "discord-akairo";
-import { GuildMember, Message } from "discord.js";
+import { DMChannel, GuildMember, Message } from "discord.js";
 import { KauriCommand } from "../../lib/commands/KauriCommand";
 import { Roles } from "../../util/constants";
 
@@ -37,7 +37,10 @@ export default class PruneCommand extends KauriCommand {
     }
 
     public async exec(message: Message, { count, user }: CommandArgs) {
+        if(message.channel instanceof DMChannel) return;
+
         const toDelete = user ? message.channel.messages.cache.filter(m => !!m.author && m.author.id === user.id) : count;
+        
         try {
             const deleted = await message.channel.bulkDelete(toDelete, true);
             this.client.logger.prune(message, deleted.size);
